+++ R/CFDataset.R
@@ -1,13 +1,14 @@
 #' CF data set
 #'
 #' @description This class represents a CF data set, the object that
-#' encapsulates a netCDF resource. You should never have to instantiate this
-#' class directly; instead, call [open_ncdf()] which will return an instance
-#' that has all properties read from the netCDF resource. Class methods can then
-#' be called, or the base R functions called with this instance.
+#'   encapsulates a netCDF resource. You should never instantiate this class
+#'   directly; instead, call [open_ncdf()] which will return an instance that
+#'   has all properties read from the netCDF resource, or [create_ncdf()] for a
+#'   new, empty instance. Class methods can then be called, or the base R
+#'   functions called with this instance.
 #'
-#' The CF data set instance provides access to all the objects in the
-#' netCDF resource, organized in groups.
+#'   The CF data set instance provides access to all the objects in the netCDF
+#'   resource, organized in groups.
 #'
 #' @export
 #' @docType class
@@ -35,19 +36,26 @@ CFDataset <- R6::R6Class("CFDataset",
     #'   CORDEX climate projection data.
     file_type = "Generic netCDF data",
 
-    #' @description Create an instance of this class.
-    #' @param resource An instance of `CFResource` that links to the netCDF
-    #' resource.
+    #' @description Create an instance of this class. Do not instantiate this
+    #'   class directly; instead, call [open_ncdf()] which will return an
+    #'   instance that has all properties read from the netCDF resource, or
+    #'   [create_ncdf()] for a new, empty instance.
+    #' @param resource An instance of `NCResource` that links to the netCDF
+    #'   resource, or a character string with the name of a new data set.
     #' @param format Character string with the format of the netCDF resource as
-    #' reported by the call opening the resource.
+    #'   reported by the call opening the resource. Ignored when argument
+    #'   `resource` is a character string.
     initialize = function(resource, format) {
-      uri <- resource$uri
-      self$name <- regmatches(uri, regexec("([^/]*)\\.nc$", uri))[[1L]][2L]
-      if (is.na(self$name))
-        self$name <- regmatches(uri, regexec("([^/]*)$", uri))[[1L]][2L]
+      if (inherits(resource, "NCResource")) {
+        uri <- resource$uri
+        self$name <- regmatches(uri, regexec("([^/]*)\\.nc$", uri))[[1L]][2L]
+        if (is.na(self$name))
+          self$name <- regmatches(uri, regexec("([^/]*)$", uri))[[1L]][2L]
 
-      private$.res <- resource
-      private$.format <- format
+        private$.res <- resource
+        private$.format <- format
+      } else
+        self$name <- resource
     },
 
     #' @description Summary of the data set printed to the console.
@@ -55,7 +63,7 @@ CFDataset <- R6::R6Class("CFDataset",
     #' is `width = ` to indicate a maximum width of attribute columns.
     print = function(...) {
       cat("<Dataset>", self$name, "\n")
-      cat("Resource   :", private$.res$uri, "\n")
+      cat("Resource   :", private$.res$uri %||% "(virtual)", "\n")
       cat("Format     :", private$.format, "\n")
       cat("Collection :", self$file_type, "\n")
       cat("Conventions:", self$conventions, "\n")
@@ -78,7 +86,7 @@ CFDataset <- R6::R6Class("CFDataset",
           cat("\nExternal variable", if (length(ev) > 1L) "s", ": ", paste(names(ev), collapse = ", "), "\n", sep = "")
       }
 
-      self$root$print_attributes(...)
+      self$root$print_attributes(width = 50L)
     },
 
     #' @description Print the group hierarchy to the console.
@@ -154,10 +162,8 @@ CFDataset <- R6::R6Class("CFDataset",
 
     #' @description List all the attributes of a group. This method returns a
     #'   `data.frame` containing all the attributes of the indicated `group`.
-    #'
     #' @param group The name of the group whose attributes to return. If the
     #'   argument is missing, the global attributes will be returned.
-    #'
     #' @return A `data.frame` of attributes.
     attributes = function(group) {
       if (missing(group))
@@ -180,6 +186,115 @@ CFDataset <- R6::R6Class("CFDataset",
     #'   value of argument `att` `NA` is returned.
     attribute = function(att, field = "value") {
       self$root$attribute(att, field)
+    },
+
+    #' @description Add an attribute to the global attributes. If an attribute
+    #'   `name` already exists, it will be overwritten.
+    #' @param name The name of the attribute. The name must begin with a letter
+    #'   and be composed of letters, digits, and underscores, with a maximum
+    #'   length of 255 characters. UTF-8 characters are not supported in
+    #'   attribute names.
+    #' @param type The type of the attribute, as a string value of a netCDF data
+    #'   type.
+    #' @param value The value of the attribute. This can be of any supported
+    #'   type, including a vector or list of values. Matrices, arrays and like
+    #'   compound data structures should be stored as a data variable, not as an
+    #'   attribute and they are thus not allowed. In general, an attribute
+    #'   should be a character value, a numeric value, a logical value, or a
+    #'   short vector or list of any of these. Values passed in a list will be
+    #'   coerced to their common mode.
+    #' @return Self, invisibly.
+    set_attribute = function(name, type, value) {
+      self$root$set_attribute(name, type, value)
+      invisible(self)
+    },
+
+    #' @description Append the text value of a global attribute. If an attribute
+    #'   `name` already exists, the `value` will be appended to the existing
+    #'   value of the attribute. If the attribute `name` does not exist it will
+    #'   be created. The attribute must be of "NC_CHAR" or "NC_STRING" type; in
+    #'   the latter case having only a single string value.
+    #' @param name The name of the attribute. The name must begin with a letter
+    #'   and be composed of letters, digits, and underscores, with a maximum
+    #'   length of 255 characters. UTF-8 characters are not supported in
+    #'   attribute names.
+    #' @param value The character value of the attribute to append. This must be
+    #'   a character string.
+    #' @param sep The separator to use. Default is `"; "`.
+    #' @param prepend Logical to flag if the supplied `value` should be placed
+    #'   before the existing value. Default is `FALSE`.
+    #' @return Self, invisibly.
+    append_attribute = function(name, value, sep = "; ", prepend = FALSE) {
+      self$root$append_attribute(name, value, sep, prepend)
+      invisible(self)
+    },
+
+    #' @description Delete attributes. If an attribute `name` is not present
+    #' this method simply returns.
+    #' @param name Vector of names of the attributes to delete.
+    #' @return Self, invisibly.
+    delete_attribute = function(name) {
+      private$.attributes <- private$.attributes[!private$.attributes$name %in% name, ]
+      invisible(self)
+    },
+
+    #' @description Add a [CFVariable] object to the data set. If there is
+    #'   another object with the same name in the group where the data variable
+    #'   should be placed an error is thrown. For objects associated with the
+    #'   data variable (such as axes, CRS, boundary variables, etc), if another
+    #'   object with the same name is otherwise identical to the associated
+    #'   object then that object will be linked from the variable, otherwise an
+    #'   error is thrown.
+    #' @param var An instance of `CFVariable` or any of its descendants.
+    #' @param group Optional. An instance of [CFGroup] where the data variable
+    #'   should be located. If omitted, the data variable will be stored in the
+    #'   root group.
+    #' @param locations Optional. A `list` whose named elements correspond to
+    #'   the names of objects associated with the data variable in argument
+    #'   `var`. Each list element has a single character string indicating the
+    #'   group in the hierarchy where the object should be stored. As an
+    #'   example, if the data variable has axes "lon" and "lat" and they should
+    #'   be stored in the parent group of `group`, then specify `locations =
+    #'   list(lon = "..", lat = "..")`. Locations can use absolute paths or
+    #'   relative paths from the `group`. Associated objects that are not in the
+    #'   list will be stored in `group`. If the argument `locations` is not
+    #'   provided, all associated objects will be stored in `group`.
+    #' @return Argument `var`, invisibly.
+    add_variable = function(var, group, locations = list()) {
+      if (missing(group)) group <- self$root
+      else if(!inherits(group, "CFGroup"))
+        stop("Argument `group` must be a CFGroup object.", call. = FALSE)
+      if (!inherits(var, "CFVariable"))
+        stop("Argument `var` must be a CFVariable object.", call. = FALSE)
+
+      group$add_variable(var, locations)
+      invisible(var)
+    },
+
+    #' @description Save the data set to file, including its subordinate objects
+    #'   such as attributes, data variables, axes, CRS, etc.
+    #' @param fn Fully-qualified file name indicating where to save the data set
+    #'   to. This argument must be provided if the data set is virtual. If the
+    #'   argument is provided on a data set that was read from a netCDF file, a
+    #'   new netCDF file will be written to the indicated location. If the
+    #'   argument is missing, the existing netCDF file will be updated.
+    #' @param pack Optional. Logical to indicate if the data should be packed;
+    #'   default is `FALSE`. Packing is only useful for numeric data; packing is
+    #'   not performed on integer values. Packing is always to the "NC_SHORT"
+    #'   data type, i.e. 16-bits per value.
+    #' @return Self, invisibly.
+    save = function(fn, pack = FALSE) {
+      # Create the netCDF file, if necessary
+      if (!missing(fn)) {
+        res <- NCResource$new(fn, write = TRUE)
+        res$create()
+        private$.res <- res
+      }
+
+      self$root$write(recursive = TRUE)
+      self$root$write_variables(pack, recursive = TRUE)
+      private$.res$close()
+      invisible(self)
     }
   ),
   active = list(
@@ -199,7 +314,7 @@ CFDataset <- R6::R6Class("CFDataset",
     #' @field uri (read-only) The connection string to the netCDF resource.
     uri = function(value) {
       if (missing(value))
-        private$.res$uri
+        if (is.null(private$.res)) NULL else private$.res$uri
     },
 
     #' @field conventions (read-only) Returns the conventions that this netCDF